requirements.txt
requests
openai
pydantic
python-dotenv